+++ src/main/java/com/anyi/reggie/entity/DishFlavor.java
@@ -32,10 +32,10 @@ public class DishFlavor implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @ApiModelProperty(value = "主键")
-    private Long id;
+    private Integer id;
 
     @ApiModelProperty(value = "菜品")
-    private Long dishId;
+    private Integer dishId;
 
     @ApiModelProperty(value = "口味名称")
     private String name;
